backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg[binary]==3.2.10
python-dotenv==1.0.1
pydantic==2.7.0
openai>=1.30.0
torch==2.2.2
